Optimize for use with std::future
This change radically simplifies mio-aio by dispensing with the notion
that an AioCb or LioCb can own its buffer.  Instead, they only operate
on slices, and the consumer is assumed to be able to track slices using
the borrow checker.  This ought to be possible using async/await, but I
haven't tested it yet.