Reduce the amount of mutex activity within AioCb
AioCb must wrap its inner nix::aio::AioCb in a mutex because:

* Operations like aio_return require a mutable reference, and so does
  registering the libc::aiocb with kqueue.
* The struct must be Send and Sync to work well with Tokio, which
  precludes options like Refcell.

But locking a mutex is slow.  Speed it up a bit by taking mutable
references in most functions that need to lock the mutex and using
Mutex::get_mut instead of Mutex::lock.  That guarantees exclusive access
at compile-time.  The main operation that still needs to lock the Mutex
is Evented::register, because mio defines that method as taking an
immutable reference.